fix eslint file to properly lint
