+++ client/screens/AllGroups/Attended.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View, Button, TouchableOpacity } from 'react-native';
+import {
+  StyleSheet, Text, View, Button, TouchableOpacity,
+} from 'react-native';
 
 const styles = StyleSheet.create({
   container: {
@@ -12,21 +14,22 @@ const styles = StyleSheet.create({
 });
 
 const Attended = ({ navigation }) => {
-
   return (
     <View style={styles.container}>
       <Text>I Am Attended</Text>
       <TouchableOpacity
-      title="Upcoming"
-      onPress={() =>
-        navigation.navigate('Upcoming', {name: 'Upcoming'})
-      } ><Text>Upcoming</Text></TouchableOpacity>
+        title="Upcoming"
+        onPress={() => navigation.navigate('Upcoming', { name: 'Upcoming' })}
+      >
+        <Text>Upcoming</Text>
+      </TouchableOpacity>
 
       <TouchableOpacity
-      title="Create"
-      onPress={() =>
-        navigation.navigate('Create', {name: 'Create'})
-      } ><Text>Create</Text></TouchableOpacity>
+        title="Create"
+        onPress={() => navigation.navigate('Create', { name: 'Create' })}
+      >
+        <Text>Create</Text>
+      </TouchableOpacity>
 
       {/* <Button
       title="Upcoming"
