+++ client/screens/AllGroups/Create.jsx
@@ -1,6 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View, SafeAreaView } from 'react-native';
+import {
+  StyleSheet, Text, View, SafeAreaView,
+} from 'react-native';
 import { Form, FormItem, Picker } from 'react-native-form-component';
 
 const styles = StyleSheet.create({
@@ -32,7 +34,7 @@ const styles = StyleSheet.create({
     fontFamily: 'Arial',
     fontWeight: 'light',
     fontSize: 18,
-  }
+  },
 });
 
 const Create = () => {
@@ -43,48 +45,55 @@ const Create = () => {
     <SafeAreaView style={styles.container}>
       {/* <Text>I Am Create</Text> */}
       <StatusBar style="auto" />
-      <Text id="create-group" style={styles.title}>CREATE GROUP</Text>
+      <Text id="create-group" style={styles.title}>
+        CREATE GROUP
+      </Text>
       <Form
         onButtonPress={() => console.warn('do something')}
         buttonText="CREATE GROUP"
         buttonStyle={styles.btnStyle}
         buttonTextStyle={styles.btnTextStyle}
       >
-        <FormItem style={styles.formInput}
+        <FormItem
+          style={styles.formInput}
           id="group-name"
           placeholder="GROUP NAME"
           isRequired
         />
-        <FormItem style={styles.formInput}
+        <FormItem
+          style={styles.formInput}
           id="group-description"
           placeholder="GROUP DESCRIPTION"
           isRequired
         />
-        <Picker style={styles.formInput}
+        <Picker
+          style={styles.formInput}
           id="select-size"
           placeholder="SELECT SIZE"
           items={[
-          { label: 'Small (0-5)', value: 'small' },
-          { label: 'Medium (6-10)', value: 'medium' },
-          { label: 'Large (11-20)', value: 'large' },
+            { label: 'Small (0-5)', value: 'small' },
+            { label: 'Medium (6-10)', value: 'medium' },
+            { label: 'Large (11-20)', value: 'large' },
           ]}
           // label="Pick a number"
           selectedValue={size}
           onSelection={(item) => setSize(item.value)}
         />
-        <Picker style={styles.formInput}
+        <Picker
+          style={styles.formInput}
           id="select-vibe"
           placeholder="SELECT VIBE"
           items={[
-          { label: 'Low', value: 'low' },
-          { label: 'Medium', value: 'medium' },
-          { label: 'High', value: 'high' },
+            { label: 'Low', value: 'low' },
+            { label: 'Medium', value: 'medium' },
+            { label: 'High', value: 'high' },
           ]}
           // label="Pick a number"
           selectedValue={vibe}
           onSelection={(item) => setVibe(item.value)}
         />
-        <FormItem style={styles.formInput}
+        <FormItem
+          style={styles.formInput}
           id="upload-img"
           placeholder="UPLOAD IMAGE"
           isRequired
